+++ .storybook/manager.tsx
@@ -1,9 +1,8 @@
 import * as React from "react";
-import {Button, TooltipLinkList, WithTooltip} from "storybook/internal/components";
-import {addons, types, useGlobals} from "storybook/manager-api";
+import {addons, types} from "storybook/manager-api";
 import theme from "./theme";
 
-import PerseusFeatureFlags from "../packages/perseus-core/src/feature-flags";
+import {FeatureFlagsToolbar} from "./feature-flags-toolbar";
 
 addons.setConfig({
     sidebar: {
@@ -14,72 +13,7 @@ addons.setConfig({
     theme: theme,
 });
 
-const TOOL_ID = "perseus/feature-flags/tool";
-
-function FeatureFlagsToolbar() {
-    const [globals, updateGlobals] = useGlobals();
-    const activeFlags: string[] = globals.featureFlags ?? [];
-
-    const toggleFlag = (flag: string) => {
-        updateGlobals({
-            featureFlags: activeFlags.includes(flag)
-                ? activeFlags.filter((f) => f !== flag)
-                : [...activeFlags, flag],
-        });
-    };
-
-    const links = [...PerseusFeatureFlags].map((flag) => {
-        const isActive = activeFlags.includes(flag);
-        return {
-            id: flag,
-            title: (
-                <span style={{display: "flex", alignItems: "center", gap: 6}}>
-                    <span
-                        aria-hidden="true"
-                        style={{
-                            width: 14,
-                            textAlign: "center",
-                            flexShrink: 0,
-                            fontWeight: "bold",
-                        }}
-                    >
-                        {isActive ? "✓" : ""}
-                    </span>
-                    {flag}
-                </span>
-            ),
-            active: isActive,
-            onClick: () => toggleFlag(flag),
-        };
-    });
-
-    const label =
-        activeFlags.length === 0
-            ? "Feature Flags"
-            : `Feature Flags: ${activeFlags.length} on`;
-
-    return (
-        <React.Suspense fallback={null}>
-            <WithTooltip
-                placement="bottom"
-                trigger="click"
-                closeOnOutsideClick
-                tooltip={<TooltipLinkList links={links} />}
-            >
-                <Button
-                    size="small"
-                    variant="outline"
-                    active={activeFlags.length > 0}
-                    title="Feature Flags"
-                >
-                    {label}
-                </Button>
-            </WithTooltip>
-        </React.Suspense>
-    );
-}
-
-addons.add(TOOL_ID, {
+addons.add("perseus/feature-flags/tool", {
     type: types.TOOL,
     title: "Feature Flags",
     match: () => true,
